+++ CHANGELOG.md
@@ -7,4 +7,4 @@
  -
 
 ## Fixes
- - 
+ - Fix rendering audio when speed is less than 0.5x